[update]版本号v3.0.3,更新SDK层(ldlidar_driver)。 1. 修复32位系统下日志打印时间戳uint64_t数据格式不对问题 2. 更改获取激光雷达扫描频率接口命名

+++ ldlidar_driver/include/logger/log_module.h
@@ -145,12 +145,12 @@ private:
 
   std::string GetCurrentTime();
 
-  inline int64_t GetCurrentLocalTimeStamp() {
+  inline uint64_t GetCurrentLocalTimeStamp() {
     //// 获取系统时间戳
     std::chrono::time_point<std::chrono::system_clock, std::chrono::nanoseconds> tp = 
       std::chrono::time_point_cast<std::chrono::nanoseconds>(std::chrono::system_clock::now());
     auto tmp = std::chrono::duration_cast<std::chrono::nanoseconds>(tp.time_since_epoch());
-    return tmp.count();
+    return (uint64_t)tmp.count();
   }
 
   std::string GetFormatValue(std::string str_value);
